tinyalsa/port config: better handling of non-supported formats
Invalid (not-tinyalsa-supported) format values were incorrectly handled:
The conversion table definition was using a "magic" value that was defined in
the body of a file (hence, it was a non-local static definition) and was
relying on another non-local static definition. Since the relative order of
those are not defined, the conversion table may or may not have been defined as
expected.

Issue: GMINL-1696
Change-Id: I07914a2fb751d69cabb03a0dea0b50692e9e4da5
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: https://android.intel.com/279642
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Beatty, Robert M <robert.m.beatty@intel.com>
Tested-by: Beatty, Robert M <robert.m.beatty@intel.com>